backend/pyproject.toml
[project]
name = "stockpilot-oms-backend"
version = "0.1.0"
description = "FastAPI backend for StockPilot OMS"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic-settings>=2.5.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
